Refactor certificates service as sagas
