+++ src/actions.js
@@ -3,7 +3,7 @@ export const ABOUT_DIALOG_DISMISSED = 'about-dialog/dismissed';
 export const ABOUT_DIALOG_TOGGLE_UPDATE_ON_START = 'about-dialog/toggle-update-on-start';
 export const ADD_SERVER_VIEW_SERVER_ADDED = 'add-server/view-server-added';
 export const CERTIFICATE_TRUST_REQUESTED = 'certificate/trust-requested';
-export const CERTIFICATES_CHANGED = 'certificates/changed';
+export const CERTIFICATES_UPDATED = 'certificates/updated';
 export const CERTIFICATES_CLEARED = 'certificates/cleared';
 export const DEEP_LINK_TRIGGERED = 'deep-link/triggered';
 export const LOADING_ERROR_VIEW_RELOAD_SERVER_CLICKED = 'loading-error-view/reload-server-clicked';
