+++ src/components/AddServerView.js
@@ -5,7 +5,7 @@ import { takeEvery } from 'redux-saga/effects';
 
 import {
 	ADD_SERVER_VIEW_SERVER_ADDED,
-	CERTIFICATES_CHANGED,
+	CERTIFICATES_UPDATED,
 } from '../actions';
 import { useSaga } from './SagaMiddlewareProvider';
 import { useServerValidation } from './ServersProvider';
@@ -76,7 +76,7 @@ export function AddServerView({
 	};
 
 	useSaga(function *() {
-		yield takeEvery(CERTIFICATES_CHANGED, function *() {
+		yield takeEvery(CERTIFICATES_UPDATED, function *() {
 			validateServerUrl(input.trim());
 		});
 	}, [input, validator]);
