+++ src/components/CertificatesProvider/index.js
@@ -1,0 +1,8 @@
+import { certificatesSaga } from './sagas';
+import { useSaga } from '../SagaMiddlewareProvider';
+
+export function CertificatesProvider({ children }) {
+	useSaga(certificatesSaga, []);
+
+	return children;
+}
